+++ test/unit/fuzz-tools.test.js
@@ -505,6 +505,7 @@ import * as wayneIndexV527 from '../../lib/wayne-index-v527.js';
 import * as mestCV528 from '../../lib/mest-c-v528.js';
 import * as thwaitesV529 from '../../lib/thwaites-v529.js';
 import * as vesikariV530 from '../../lib/vesikari-v530.js';
+import * as ehitV531 from '../../lib/ehit-v531.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -955,6 +956,7 @@ const MODULES = {
   'mest-c-v528.js': mestCV528,
   'thwaites-v529.js': thwaitesV529,
   'vesikari-v530.js': vesikariV530,
+  'ehit-v531.js': ehitV531,
   'tb-testing.js': tbTesting,
 };